requirements.txt
openpyxl>=3.1.0
anthropic>=0.40.0
langgraph>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.28.0
